+++ common/predictive-text/build.sh
@@ -28,7 +28,6 @@ WORKER_OUTPUT=build/intermediate
 INCLUDES_OUTPUT=build/includes
 NAKED_WORKER=$WORKER_OUTPUT/index.js
 EMBEDDED_WORKER=$WORKER_OUTPUT/embedded_worker.js
-LEXICAL_MODELS_TYPES=../lexical-model-types
 
 # Builds the top-level JavaScript file for use in browsers (the second stage of compilation)
 build-browser () {
@@ -171,10 +170,6 @@ done
 
 # Check if Node.JS/npm is installed.
 verify_npm_setup $fetch_deps
-
-pushd ../core/web/utils/src
-./build.sh -skip-package-install || fail "Utility library package compilation failed."
-popd
 
 # Ensure that the build-product destination for any generated include .d.ts files exists.
 if ! [ -d $INCLUDES_OUTPUT ]; then